Rename FIDData and SpectroData(::MRExperiment) constructor
The rename here is to be more explicit that FIDData is not merely a
container of FIDs, but a multidimensional spectroscopy experiment.

Also utilize YAPS metadata parsing to fully extract parameters so we can
use our own loader rather than suspect.py for reading twix files.  The
julia version of this parser is about 10x faster than the python one.