Start new MR data recognizer infrastructure
Start reorganizing code in attempt to silo off Siemens format-specific
stuff from the processing code.

The idea here is to allow an experiment-specific type to be overlaid
onto the raw MR acquisition data so that processing code just needs to
deal with the layout, while a separate section of the code can deal with
recognizing manufacturer-specific stuff. This will also help in the
future if we want to split the Siemens-specific stuff out into its own
repo.

* mr_recognize() function, to recognize the type of an MRExperiment
  based on the metadata.
* LCOSY data layout definition struct, the first to be used
* Basic recognition of srcosy and svs_lcosy sequences for TRI twix data